+++ docker/.dockerignore
@@ -1,5 +1,5 @@
 # Gradle 관련
-.gradle/
+../.gradle/
 build/
 !gradle/wrapper/gradle-wrapper.jar
