+++ src/main/java/com/be/sportizebe/domain/chat/dto/request/ChatRoomCreateRequest.java
@@ -1,4 +1,4 @@
-package com.be.sportizebe.domain.chat.dto;
+package com.be.sportizebe.domain.chat.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
