+++ src/main/java/com/be/sportizebe/domain/chat/dto/response/ChatMessageResponse.java
@@ -1,4 +1,4 @@
-package com.be.sportizebe.domain.chat.dto;
+package com.be.sportizebe.domain.chat.dto.response;
 
 import com.be.sportizebe.domain.chat.entity.ChatMessage;
 import lombok.Builder;
